modification couleur

+++ fractol/includes/fractol.h
@@ -6,7 +6,7 @@
 /*   By: yvillepo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/01 01:56:28 by yvillepo          #+#    #+#             */
-/*   Updated: 2018/01/19 04:27:04 by yvillepo         ###   ########.fr       */
+/*   Updated: 2018/01/19 05:26:25 by yvillepo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,6 +79,6 @@ void				affiche_mlx(t_mlx *mlx);
 void				mandelbrot_image1(t_mlx *mlx,t_complex *c1,t_complex *c2);
 void				mlx_free(t_mlx **m);
 unsigned int					color(float percent);
-unsigned int		color1(float percent);
+unsigned int		color1(double percent);
 
 #endif
